refactor: tsx rename
